requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.1
pydantic==2.11.5
google-genai==1.16.1